Use explicit dtypes to work around varying default types on the platforms

+++ python/turbodbc_test/test_executemanycolumns.py
@@ -9,7 +9,7 @@ from helpers import open_cursor, for_each_database
 def test_integer_column(dsn, configuration):
     with open_cursor(configuration) as cursor:
         with query_fixture(cursor, configuration, 'INSERT INTEGER') as table_name:
-            columns = [array([17, 23, 42])]
+            columns = [array([17, 23, 42], dtype='int64')]
             cursor.executemanycolumns("INSERT INTO {} VALUES (?)".format(table_name), columns)
 
             results = cursor.execute("SELECT A FROM {} ORDER BY A".format(table_name)).fetchall()
@@ -20,7 +20,7 @@ def test_integer_column(dsn, configuration):
 def test_integer_column_exceeds_buffer_size(dsn, configuration):
     with open_cursor(configuration, parameter_sets_to_buffer=2) as cursor:
         with query_fixture(cursor, configuration, 'INSERT INTEGER') as table_name:
-            columns = [array([17, 23, 42])]
+            columns = [array([17, 23, 42], dtype='int64')]
             cursor.executemanycolumns("INSERT INTO {} VALUES (?)".format(table_name), columns)
 
             results = cursor.execute("SELECT A FROM {} ORDER BY A".format(table_name)).fetchall()
@@ -31,7 +31,7 @@ def test_integer_column_exceeds_buffer_size(dsn, configuration):
 def test_masked_integer_column(dsn, configuration):
     with open_cursor(configuration) as cursor:
         with query_fixture(cursor, configuration, 'INSERT INTEGER') as table_name:
-            columns = [MaskedArray([17, 23, 42], mask=[False, True, False])]
+            columns = [MaskedArray([17, 23, 42], mask=[False, True, False], dtype='int64')]
             cursor.executemanycolumns("INSERT INTO {} VALUES (?)".format(table_name), columns)
 
             results = cursor.execute("SELECT A FROM {} ORDER BY A".format(table_name)).fetchall()
@@ -42,7 +42,7 @@ def test_masked_integer_column(dsn, configuration):
 def test_masked_integer_column_with_shrunk_mask(dsn, configuration):
     with open_cursor(configuration) as cursor:
         with query_fixture(cursor, configuration, 'INSERT INTEGER') as table_name:
-            columns = [MaskedArray([17, 23, 42], mask=False)]
+            columns = [MaskedArray([17, 23, 42], mask=False, dtype='int64')]
             cursor.executemanycolumns("INSERT INTO {} VALUES (?)".format(table_name), columns)
 
             results = cursor.execute("SELECT A FROM {} ORDER BY A".format(table_name)).fetchall()
@@ -53,7 +53,7 @@ def test_masked_integer_column_with_shrunk_mask(dsn, configuration):
 def test_masked_integer_column_exceeds_buffer_size(dsn, configuration):
     with open_cursor(configuration, parameter_sets_to_buffer=2) as cursor:
         with query_fixture(cursor, configuration, 'INSERT INTEGER') as table_name:
-            columns = [MaskedArray([17, 23, 42], mask=[True, False, True])]
+            columns = [MaskedArray([17, 23, 42], mask=[True, False, True], dtype='int64')]
             cursor.executemanycolumns("INSERT INTO {} VALUES (?)".format(table_name), columns)
 
             results = cursor.execute("SELECT A FROM {} ORDER BY A".format(table_name)).fetchall()
@@ -64,7 +64,7 @@ def test_masked_integer_column_exceeds_buffer_size(dsn, configuration):
 def test_multiple_columns(dsn, configuration):
     with open_cursor(configuration) as cursor:
         with query_fixture(cursor, configuration, 'INSERT TWO INTEGER COLUMNS') as table_name:
-            columns = [array([17, 23, 42]), array([3, 2, 1])]
+            columns = [array([17, 23, 42], dtype='int64'), array([3, 2, 1], dtype='int64')]
             cursor.executemanycolumns("INSERT INTO {} VALUES (?, ?)".format(table_name), columns)
 
             results = cursor.execute("SELECT A, B FROM {} ORDER BY A".format(table_name)).fetchall()
